Relocate all calls of UsedSettingRepository
UsedSettingRepository has been moved to MinterService as its more
appropriate for the MinterService to update the database.
